+++ kwcoco/metrics/voc_metrics.py
@@ -11,7 +11,7 @@ class VOC_Metrics(ub.NiceRepr):
     :func:`VOC_Metrics.score` function.
 
     Attributes:
-        recs (Dict[int, List[dict]):
+        recs (Dict[int, List[dict]]):
             true boxes for each image.  maps image ids to a list of records
             within that image.  Each record is a tlbr bbox, a difficult flag,
             and a class name.
@@ -20,6 +20,9 @@ class VOC_Metrics(ub.NiceRepr):
             VOC formatted prediction preditions.  mapping from class index to
             all predictions for that category.  Each "line" is a list of
             [[<imgid>, <score>, <tl_x>, <tl_y>, <br_x>, <br_y>]].
+
+        classes (None | List[str] | kwcoco.CategoryTree):
+            class names
     """
     def __init__(self, classes=None):
         self.recs = {}
@@ -265,7 +268,7 @@ def _voc_eval(lines, recs, classname, iou_thresh=0.5, method='voc2012',
     for imagename in imagenames:
         R = [obj for obj in recs2[imagename] if obj['name'] == classname]
         bbox = np.array([x['bbox'] for x in R])
-        difficult = np.array([x['difficult'] for x in R]).astype(np.bool)
+        difficult = np.array([x['difficult'] for x in R]).astype(bool)
         det = [False] * len(R)
         npos = npos + sum(~difficult)
         class_recs[imagename] = {'bbox': bbox,