+++ CHANGELOG.md
@@ -5,7 +5,18 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.5.2 - Unreleased
+## Version 0.5.3 - Unreleased
+
+### Added
+* Add `role` property for annotations in the schema.
+
+### Fixed
+* Added `instance_id` for mmlab tracking
+* Fixed np.bool
+* Fixed np.float
+
+
+## Version 0.5.2 - Released 2022-12-15
 
 ### Added